Removes unused parameters and puts the classes into their own files
