+++ src/main/java/server/ExecutorServiceFactory.java
@@ -1,0 +1,7 @@
+package server;
+
+import java.util.concurrent.ExecutorService;
+
+public interface ExecutorServiceFactory {
+    ExecutorService create();
+}
