+++ src/main/java/server/FixedThreadPoolExecutorService.java
@@ -1,0 +1,17 @@
+package server;
+
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+
+public class FixedThreadPoolExecutorService implements ExecutorServiceFactory {
+    private final int numberOfThreads;
+
+    public FixedThreadPoolExecutorService(int numberOfThreads) {
+        this.numberOfThreads = numberOfThreads;
+    }
+
+    @Override
+    public ExecutorService create() {
+        return Executors.newFixedThreadPool(numberOfThreads);
+    }
+}
