+++ src/main/java/server/HttpServerRunner.java
@@ -11,18 +11,16 @@ public class HttpServerRunner {
     public static void main(String... args) throws IOException {
         CommandLineArgumentParser commandLineArgumentParser = new CommandLineArgumentParser(System.getProperty("user.home"));
 
-        String host = "localhost";
         int port = commandLineArgumentParser.extractPort(args);
         String publicDirectory = commandLineArgumentParser.extractPublicDirectory(args);
 
         HttpServerSocket httpServerSocket = new HttpServerSocket(new ServerSocket(port), new HttpResponseFormatter());
 
         HttpServer httpServer = new HttpServer(
-                host,
-                port,
                 httpServerSocket,
                 new HttpRequestParser(),
-                new HttpRouteProcessor(new FileResourceHandler(publicDirectory)), new FixedThreadPoolExecutorService(4)
+                new HttpRouteProcessor( new FileResourceHandler(publicDirectory)),
+                new FixedThreadPoolExecutorService(4)
         );
 
         start(httpServer);
