+++ src/main/java/server/RequestThreadExecutorService.java
@@ -1,0 +1,22 @@
+package server;
+
+import java.util.concurrent.ExecutorService;
+
+public class RequestThreadExecutorService implements ThreadExecutorService {
+
+    private final ExecutorService executor;
+
+    public RequestThreadExecutorService(ExecutorService executor) {
+        this.executor = executor;
+    }
+
+    @Override
+    public void execute(Runnable task) {
+        executor.execute(task);
+    }
+
+    @Override
+    public void shutdown() {
+        executor.shutdown();
+    }
+}
