+++ src/main/java/server/ThreadExecutorService.java
@@ -1,0 +1,6 @@
+package server;
+
+public interface ThreadExecutorService {
+    void execute(Runnable r);
+    void shutdown();
+}
